add contract

+++ clientlib/src/main/java/wallet/idhub/com/clientlib/interfaces/IdentityChainViewer.java
@@ -1,6 +1,8 @@
 package wallet.idhub.com.clientlib.interfaces;
 
+import java.math.BigInteger;
+
 public interface IdentityChainViewer {
-	public Listen<Long> getEIN(String associate);
+	public Listen<BigInteger> getEIN(String associate);
 	public Listen<Identity> getIdentity(long ein);
 }
